feat: first mvp iteration, now it can collect content from files recursively and has extension support
