+++ src/constants.py
@@ -1,15 +1,2 @@
-DEFAULT_OUTPUT_FILENAME = "output.txt"
-TEXT_FILE_EXTENSIONS = [
-    ".txt",
-    ".md",
-    ".py",
-    ".json",
-    ".csv",
-    ".html",
-    ".js",
-    ".ts",
-    ".css",
-    ".xml",
-    ".yml",
-    ".yaml"
-]
+ALLOWED_FILE_EXTENSIONS = [".txt", ".json", ".java", ".py"]
+OUTPUT_FILENAME = "output.txt"
